requirements.txt
Flask
opencv-python-headless
pytesseract
pillow
textblob
nltk
